create utils.js
